requirements.txt
Django>=4.2,<5.0
python-dotenv
psycopg2-binary
dj-database-url
djangorestframework
supabase
django-cors-headers
